# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DigiScore-WA API",
        "description": "Copilote d'\u00e9ligibilit\u00e9 et de plafond \u2014 institutions CIF. Contrat Front \u2194 Back (cl\u00e9s JSON FR).",
        "version": "0.3.0"
    },
    "paths": {
...
        "/demandes/{demande_id}/anomalies": {
            "get": {
                "tags": [
                    "ml"
                ],
                "summary": "Get Anomalies",
                "operationId": "get_anomalies_demandes__demande_id__anomalies_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnomaliesOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demandes/{demande_id}/simuler": {
            "post": {
                "tags": [
                    "ml"
                ],
                "summary": "Post Simulation",
                "operationId": "post_simulation_demandes__demande_id__simuler_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SimulationIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimulationOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demandes/{demande_id}/simulation/enregistrer": {
            "post": {
                "tags": [
                    "ml"
                ],
                "summary": "Post Simulation Snapshot",
                "operationId": "post_simulation_snapshot_demandes__demande_id__simulation_enregistrer_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SimulationIn"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimulationSavedOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demandes/{demande_id}/simulations": {
            "get": {
                "tags": [
                    "ml"
                ],
                "summary": "Get Simulations",
                "operationId": "get_simulations_demandes__demande_id__simulations_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SimulationSavedOut"
                                    },
                                    "title": "Response Get Simulations Demandes  Demande Id  Simulations Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/portefeuille/alertes": {
            "get": {
                "tags": [
                    "ml"
                ],
                "summary": "Get Alertes",
                "operationId": "get_alertes_portefeuille_alertes_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertesOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demandes/{demande_id}/ml/scorecard": {
            "get": {
                "tags": [
                    "ml"
                ],
                "summary": "Get Scorecard Shadow",
                "description": "Scorecard v3 en shadow : probabilite indicative, jamais decisionnelle.",
                "operationId": "get_scorecard_shadow_demandes__demande_id__ml_scorecard_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScorecardShadowOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demandes/{demande_id}/ml/plafond": {
            "get": {
                "tags": [
                    "ml"
                ],
                "summary": "Get Plafond Ml",
                "description": "Plafond ML indicatif : decote du plafond regles, null si knockout.",
                "operationId": "get_plafond_ml_demandes__demande_id__ml_plafond_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "demande_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Demande Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlafondMlOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AlertesOut": {
                "properties": {
                    "model_version": {
                        "type": "string",
                        "title": "Model Version"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/AlerteOut"
                        },
                        "type": "array",
                        "title": "Items"
                    }
                },
                "type": "object",
                "required": [
                    "model_version"
                ],
                "title": "AlertesOut"
            },
...
            "AnomaliesOut": {
                "properties": {
                    "model_version": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Version"
                    },
                    "scope_excluded": {
                        "type": "boolean",
                        "title": "Scope Excluded",
                        "default": false
                    },
                    "anomaly_score": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Anomaly Score"
                    },
                    "anomalies": {
                        "items": {
                            "$ref": "#/components/schemas/AnomalieOut"
                        },
                        "type": "array",
                        "title": "Anomalies"
                    }
                },
                "type": "object",
                "title": "AnomaliesOut"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ScorecardShadowOut": {
                "properties": {
                    "model_version": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Version"
                    },
                    "mode": {
                        "type": "string",
                        "title": "Mode",
                        "default": "shadow"
                    },
                    "probabilite_defaut": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Probabilite Defaut"
                    },
                    "score_global_ml": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score Global Ml"
                    },
                    "niveau_risque": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Niveau Risque"
                    },
                    "regles_knockout": {
                        "type": "boolean",
                        "title": "Regles Knockout",
                        "default": false
                    },
                    "top_factors": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Top Factors"
                    },
                    "contributions": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Contributions"
                    },
                    "warning": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Warning"
                    }
                },
                "type": "object",
                "title": "ScorecardShadowOut"
            },
...
            "SimulationIn": {
                "properties": {
                    "scenario": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "additionalProperties": true,
                                "type": "object"
                            }
                        ],
                        "title": "Scenario",
                        "default": "normal"
                    },
                    "montant": {
                        "anyOf": [
                            {
                                "type": "number",
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Montant"
                    },
                    "duree_mois": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 60.0,
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Duree Mois"
                    },
                    "horizon_mois": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 24.0,
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Horizon Mois"
                    },
                    "iterations": {
                        "type": "integer",
                        "maximum": 10000.0,
                        "minimum": 1.0,
                        "title": "Iterations",
                        "default": 1000
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed",
                        "default": 42
                    }
                },
                "type": "object",
                "title": "SimulationIn"
            },
            "SimulationOut": {
                "properties": {
                    "application_id": {
                        "type": "integer",
                        "title": "Application Id"
                    },
                    "scenario": {
                        "type": "string",
                        "title": "Scenario"
                    },
                    "scenario_libelle": {
                        "type": "string",
                        "title": "Scenario Libelle",
                        "default": ""
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "default": ""
                    },
                    "mecanique": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Mecanique"
                    },
                    "model_version": {
                        "type": "string",
                        "title": "Model Version"
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed"
                    },
                    "p_incident": {
                        "type": "number",
                        "title": "P Incident"
                    },
                    "p10": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P10"
                    },
                    "p50": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P50"
                    },
                    "p90": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P90"
                    },
                    "mois_critique": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mois Critique"
                    },
                    "explication": {
                        "type": "string",
                        "title": "Explication"
                    },
                    "montant": {
                        "type": "number",
                        "title": "Montant"
                    },
                    "duree_mois": {
                        "type": "integer",
                        "title": "Duree Mois"
                    },
                    "horizon_mois": {
                        "type": "integer",
                        "title": "Horizon Mois"
                    },
                    "iterations": {
                        "type": "integer",
                        "title": "Iterations"
                    }
                },
                "type": "object",
                "required": [
                    "application_id",
                    "scenario",
                    "model_version",
                    "seed",
                    "p_incident",
                    "p10",
                    "p50",
                    "p90",
                    "explication",
                    "montant",
                    "duree_mois",
                    "horizon_mois",
                    "iterations"
                ],
                "title": "SimulationOut"
            },
            "SimulationSavedOut": {
                "properties": {
                    "application_id": {
                        "type": "integer",
                        "title": "Application Id"
                    },
                    "scenario": {
                        "type": "string",
                        "title": "Scenario"
                    },
                    "scenario_libelle": {
                        "type": "string",
                        "title": "Scenario Libelle",
                        "default": ""
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "default": ""
                    },
                    "mecanique": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Mecanique"
                    },
                    "model_version": {
                        "type": "string",
                        "title": "Model Version"
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed"
                    },
                    "p_incident": {
                        "type": "number",
                        "title": "P Incident"
                    },
                    "p10": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P10"
                    },
                    "p50": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P50"
                    },
                    "p90": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "P90"
                    },
                    "mois_critique": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mois Critique"
                    },
                    "explication": {
                        "type": "string",
                        "title": "Explication"
                    },
                    "montant": {
                        "type": "number",
                        "title": "Montant"
                    },
                    "duree_mois": {
                        "type": "integer",
                        "title": "Duree Mois"
                    },
                    "horizon_mois": {
                        "type": "integer",
                        "title": "Horizon Mois"
                    },
                    "iterations": {
                        "type": "integer",
                        "title": "Iterations"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "application_id",
                    "scenario",
                    "model_version",
                    "seed",
                    "p_incident",
                    "p10",
                    "p50",
                    "p90",
                    "explication",
                    "montant",
                    "duree_mois",
                    "horizon_mois",
                    "iterations",
                    "id",
                    "created_at"
                ],
                "title": "SimulationSavedOut"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}